+++ files/root/index.php
@@ -2,7 +2,7 @@
 /**
  * Root index file for DDEV WP Local Setup.
  *
- * @package CassidyDC/DDEV-WP-Local-Setup
+ * @package CassidyDC/ddev-local-wp-setup
  */
 
 // Load the WordPress index file.